file item

+++ item.h
@@ -11,7 +11,7 @@ typedef struct Item {
     char *name;
     int hpMax;
     int shield;
-    int dmg;
+    float dmg;
     enum boolean ps;         //bouléen true=1 ou false=0
     enum boolean ss;         //bouléen
     enum boolean flight;     //bouléen
@@ -19,7 +19,7 @@ typedef struct Item {
 
 
 
-Item * newItem(char *name, int hpMax, int shield, int dmg, boolean ps, boolean ss, boolean flight);
+Item * newItem(char *name, int hpMax, int shield, float dmg, boolean ps, boolean ss, boolean flight);
 void freeItem(Item *item);
 void showItem(Item item);
 void printItem(Item item, FILE *f);
